feat(optimization): moves main to bootstrap script
